+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Removed sandbox mode from macOS
+- Added `sudoPassword` to `UserChoice` model
+- Added asking for Sudo Password on macOS only in `CustomizeStepView`
 - macOS Script:
   - Install Flutter SDK
   - Unzip Flutter SDK
@@ -17,11 +18,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Install Git (requires `brew`)
   - Install Android Studio
   - Install IntelliJ IDEA
-  <!-- - Install Visual Studio Code -->
+  - Install Visual Studio Code
 
 ### Fixed
 
 - Naming mistake in `CustomizeView`
+- Borders of text fields in `CustomizeView`
+
+### Removed
+
+- Removed sandbox mode from macOS
 
 ## [0.0.2] - 2020-07-23